+++ scanner/__init__.py
@@ -10,6 +10,7 @@ from scanner.entropy import shannon_entropy
 from scanner.filters import is_placeholder
 from scanner.file_handler import ScanConfig
 from scanner.git_mode import GitError
+from scanner.ignore import IgnoreError
 from scanner.models import PatternMatch, ScanResult, SecretFinding, SecretPattern, Severity
 from scanner.patterns import PatternEngine
 from scanner.scanner import Scanner
@@ -22,6 +23,7 @@ __all__ = [
     "Detector",
     "FileScan",
     "GitError",
+    "IgnoreError",
     "is_placeholder",
     "is_sensitive_identifier",
     "shannon_entropy",